Move EvaluationReport types to own namespace and folder
