+++ src/FaluSdk/FaluClient.cs
@@ -1,10 +1,11 @@
-﻿using Falu.Evaluations;
+﻿using Falu.EvaluationReports;
+using Falu.Evaluations;
 using Falu.Events;
 using Falu.FileLinks;
 using Falu.Files;
 using Falu.Identity;
-using Falu.IdentityVerifications;
 using Falu.IdentityVerificationReports;
+using Falu.IdentityVerifications;
 using Falu.Messages;
 using Falu.MessageStreams;
 using Falu.MessageTemplates;
