+++ tests/FaluSdk.Tests/Clients/EvaluationsServiceClientTests.cs
@@ -3,7 +3,6 @@ using Falu.Evaluations;
 using System.Net;
 using System.Net.Mime;
 using System.Text;
-using System.Text.Json;
 using Tingle.Extensions.JsonPatch;
 using Xunit;
